Main Menu + no. player select
Added a main menu separate level, with placeholders for credits, options and quit buttons. The main menu also has a "Start" button which, when clicked, displays a screen of four buttons that provide the player with a choice of how many players they want to have. The information of how many players is stored in a struct within the game instance so it can be transferred to other levels through automatic saving and loading.