Created a sub-folder to share the widgets between myself and Elliot that allows the callout of other players depending on which player presses the key to gain control of the mouse. Will begin to check to see which player has managed to press the key so they can call out the correct player along assistance with Elliot.
Added a key-binding check that allows mouse control on all four players, as well as insuring the correct key is pressed no matter how many players are in the scene due to the player index. Removes the viewport UI prompt to press the key within 5 seconds, although beginning to see if it can be called properly or not. Discussed with Elliot further ideas that we needed and documented what we need to do next with the plans and documented them in comment boxes in the level scene/on Trello.