FINAL COMMIT
Added a synopsis screen to the main menu so the players now know what to expect from the game without it being confusing. Added sound effects into the game that play either when the player bounces or as part of the background after a random time between 2-20 seconds. Cleaned up existing/added new folders with assets I created/had been working on in the project. Bug fixing (if needed).